feat: option to provide bearer token on manager requests
When developing locally towards a manager that is running in
Open Source Cloud we need to provide the service access token
as a bearer token. This change adds the option to set the
environment variable VITE_BACKEND_API_KEY with the service access
token and also included instructions on how to obtain it.

When application is deployed in the OSC platform this is not
needed as authorization is performed using cookies.